Create therm.py
This program uses the DS18B20 waterproof temperature sensors from adafruit. The sensors use Dallas Protocol.
This is dsigned to be used with a Heat Transfer experiment. It updates a graph as new data comes in. It also writes the 
data to a text file that can be analyzed after.